Move tools code and add a .gitignore
The tool mkapertemplate has a Makefile rule to build it.
This rule is not invoked by default when building assess.
It is mentioned in comments inside defs.h.

Other code relating to pi_term and the runs_table were
moved into the tools directory to be consisent.  Comments
mentioning these files were updated to reflect these moves.