Improve longestRunOfOnes.c and rank.c
The comments in longestRunOfOnes.c and in rank.c have been improved.
Where possible, the code has been made more readable and the steps
in the iterate function have been organized to match the paper.

In the rank test, the precondition check has been fixed and moved
to the init function, since the only things that have to be checked in
this test can be tested before the iteration.