Improve universal, overlapping and other tests
In the the universal test:
A magic number representing the maximum sequence length
has been removed.
The method used to compute L has been fixed to work correctly
in all cases and documented to be easier to understand.
A bug in the computation of K has been fixed.
The documentation has been considerably improved.

In the overlapping test:
The documentation has been improved. A useless legacy variable
and a function have been removed.

In the nonOverlapping test:
Some checks have been reintroduced and their behavior has been
made more consistent with the other tests.

In the frequency and discreteFourierTransform tests:
An optimization has been made in the way the sequence is
adjusted from 0s and 1s to -1s and +1s.

Other tests:
Minor code style improvements have been made here and there.