Improve defs.h and linearComplexity.c
The upperbound on the length of a single bit stream has been removed,
since in all the tests only lower bounds are specified.

The template size boundaries are not applied anymore to the overlapping
template matchings test, since they are never mentioned in the paper
(where they are only used on the non overlapping test).

In the linear complexity test, the record of the value of T in the v
counters has been made K dependent, and not hardcoded anymore
with the classes of K = 6.

A better comment has been added to the usage error thrown when
the given length of the bitstream is not multiple of 8.