Add -s flag (replace -n) and change number of bins
The -s flag to print stats in stats.txt and results.txt files was added.
It replaces the -n flag by doing the opposite thing.

The usage string has been fixed (it contained some wrong flags).
A bool value was added to state to note down whether the -s flag
was used.

The default number of bins is now computed dynamically when the
legacy output mode is disabled and a custom value is not added.
It is forced to the default value (ignoring custom parameters) when
the legacy output mode is enabled.