Fix -s flag and remove -Werror from CFLAGS
The '-s' flag, which was returning error on getopt() has been fixed (it
was simply missing from the getopt parameters).

The -Werror flag has been removed from the CFLAGS for the GCC compiler
in the Makefile. Please report to us if the compiler gives you any
warning in the future.