Make STS multi-threaded, fix bugs and improve code
The code of the STS now runs the iterations in parallel and by default
uses as many threads as the cores of the machine where it's running.
Several parts of the code have been changed to support multi-threads.

New checks have been added as firewalls to all the functions affected.

A new flag -T has been added to specify a custom number of threads.

Some unused #include statements have been removed.
Some unused variables / assignments have been removed.

In universal.c there was a bad issue with the value of K, which was
computed in such a way that (Q + K) * MIN_L > n. This lead to accessing
unallocated blocks of memory in the iterate function. Now the issue
has been fixed by letting K = 100 * Q = 1000 * 2^L, as it is done
in the table printed in the original STS paper and as it is done to get
the value of L. The correct formula to use remains unclear to me, but now
at least the code is consistent with the paper and does not have memory
issues.

The modes in which STS can operate have been made clearer, as well as
the usage instructions.

Some inconsistencies in the debug prints have been fixed.

The generators.c file has been improved. There was also a bug which
caused two calls to free() of epsilon when g != 0, which has been fixed.

In utilities.c, several methods have been improved / restructured for the
multi-threaded computation.

The text of the result.txt file when using an internal generator
has been improved.

The README.md file has been updated.
The version is now 3.1.0.