Add distributed testing feature
Now it is possible to run STS in multiple computers and later evaluate
the results of the individual runs at once (this will be useful when
testing huge data-sets). To this goal, two new modes of operation
have been added, as well as one new flag (-d).

Minor fixes and improvements have been done to other parts of the code.

The README and the introductory comment in sts.c have been updated.