Improve distributed feature and update constants
The default block size for the BLOCK_FREQUENCY test has been updated
to make the test execute and not be disabled all the time.

The bitcount argument has been made optional and its default value has
been set to 2^20 = 1048576.

The -d flag (for the distributed mode) now takes a directory arg, and
the .pvalues files are now taken from that directory dynamically basing
on their filename. Also, now the filenames are stored in a linked list.

The result.txt output has been improved and made easier to grep.

The README.md file has been updated.