Make batch mode the default mode
Removed -b as batch mode is now the default.

Added -A to turn on obsolete interactive mode.

Removed -p as this is no longer needed.  The -A forces
promtping for parameters.

Warned against using -P to change values.

Warned that -g 1-9 will go away in the future.

Flipped the meaning of the promptFlag boolean.