Attempt to complete metrics when missing pvalues
When -m a encounters an I/O error such as a truncated
pvalues file, instead of just aborting during the metrics phase,
the code will attempt to use the pvalues that it found.
The results may be bogus in such a case, but it may be better
to warn the user than to kill the assessment altogether.